ABANDONING TE3_hub - unfortunately, during testing several factors rule this approach out - the SGTL5000 I2S<->audio works ok and may still be in future designs, but ... - there were noticable and regular glitches in USB Audiio OUT-->IN on both Win10 and the iPad - certainly USB Midi and USB Audio on the teensy is not well supported and generates huge amounts    of noise, so I would still need an external USB Hub for the FTP - even with the glitches there were problems getting the USB Audio device to work on newer iPad,    particularly that Tonestack refused to utilize it correctly, either stand-alone or in combination    with AudioBus, although AudioBus did work reasonably well. - I don't have enough lifetime left to try to debug teensy USB Audio and/or Midi problems    particularly against a closed architecture like the iPad.
It just won't work for a pro-quality that I want.
So I'm back to the iRigHD2 as the USB Audio device, an external hub with the FTP.

The only sort of positive was that in playing with the SGTL5000 I believe it is probably
"good enough" to replace the Audio Injector Octo in the system with a fairly big
space savings and better form factor for incorporation into TE3.

Oh well.

checking final version in